# openapi.yml
openapi: 3.0.3
info:
  title: MasterFabric Go API
  description: |
    Authentication and Health Check endpoints for MasterFabric Go.
    
    ## Authentication Flow
    1. Register a new user with POST /api/v1/auth/register
    2. Login with POST /api/v1/auth/login to get access token and refresh token cookie
    3. Use access token in Authorization header for protected endpoints
    4. Refresh token is stored in HttpOnly cookie and rotated on each refresh
    5. Logout with POST /api/v1/auth/logout to revoke refresh token
    
    ## Bugs Fixed
    - Chi middleware ordering: Gateway middleware moved before WebSocket route
    - Redis nil interface check: Added reflect.ValueOf().IsNil() check in health handler
    - IP address extraction: Parse RemoteAddr to remove port for PostgreSQL inet compatibility
    - Database schema: Changed ip_address column from INET to TEXT for pgx binary format compatibility
  version: 1.0.0
  contact:
    name: MasterFabric Team

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://masterfabric-go.onrender.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User registration, login, and token management

paths:
  /health/live:
    get:
      tags: [Health]
      summary: Liveness check
      description: |
        Returns 200 if the server is alive. Use this for Kubernetes liveness probes.
      operationId: liveness
      responses:
        '200':
          description: Server is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: alive

  /health/ready:
    get:
      tags: [Health]
      summary: Readiness check
      description: |
        Checks database and cache connectivity. Returns 200 if all services are healthy.
        Use this for Kubernetes readiness probes.
      operationId: readiness
      responses:
        '200':
          description: All services healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: One or more services unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Creates a new user account. Returns the created user information.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: |
        Authenticates a user and returns an access token. A refresh token is set
        as an HttpOnly cookie for token rotation.
        
        **Cookie Security:**
        - Production (Secure=true): SameSite=None, Secure, HttpOnly
        - Development (Secure=false): SameSite=Strict, HttpOnly
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: |
                Refresh token cookie (HttpOnly, not accessible via JavaScript).
                Format: `refresh_token=<token>; Path=/api/v1/auth; HttpOnly; SameSite=None; Secure`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Rotates the refresh token and returns a new access token.
        The refresh token is sent via HttpOnly cookie (not in request body).
        
        **Token Rotation:**
        - Old refresh token is revoked
        - New refresh token is set in cookie
        - New access token is returned in response body
        
        **Replay Detection:**
        - If a reused (already rotated) token is detected, all sessions for the
          user are revoked for security.
      operationId: refresh
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              description: New refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: |
        Revokes the current refresh token and clears the cookie.
        Requires valid JWT in Authorization header.
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/logout-all:
    post:
      tags: [Authentication]
      summary: Logout from all devices
      description: |
        Revokes all refresh tokens for the current user across all devices.
        Requires valid JWT in Authorization header.
      operationId: logoutAll
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All sessions terminated
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/sessions:
    get:
      tags: [Authentication]
      summary: List active sessions
      description: |
        Returns a list of active sessions/devices for the current user.
        Requires valid JWT in Authorization header.
      operationId: listSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sessions list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionsResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /api/v1/auth/login.
        Format: `Authorization: Bearer <token>`

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: securepassword123
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: securepassword123

    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        status:
          type: string
          enum: [active, inactive, suspended]
          example: active
        created_at:
          type: string
          format: date-time
          example: 2026-01-01T00:00:00Z

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 900
        user:
          $ref: '#/components/schemas/UserInfo'

    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          description: New JWT access token
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 900

    SessionInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Session ID
        created_at:
          type: string
          format: date-time
          description: When the session was created
        ip_address:
          type: string
          description: Client IP address
          example: 192.168.1.1
        user_agent:
          type: string
          description: Client user agent string
          example: Mozilla/5.0...

    SessionsResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionInfo'

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ready, not ready]
          example: ready
        services:
          type: object
          additionalProperties:
            type: string
          example:
            postgres: healthy
            redis: healthy

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: invalid credentials